Refactor color_rgb_to_hex to fix wrong type (#38)
* Update color.py

* ds

+++ homeassistant/util/color.py
@@ -489,7 +489,7 @@ def color_rgbww_to_rgb(
     return match_max_scale((r, g, b, cw, ww), rgb)  # type: ignore[return-value]
 
 
-def color_rgb_to_hex(r: int, g: int, b: int) -> str:
+def color_rgb_to_hex(r: float, g: float, b: float) -> str:
     """Return a RGB color from a hex color string."""
     return f"{round(r):02x}{round(g):02x}{round(b):02x}"
